// types/resultType.ts
export type TResult = {
  status: string;
  traceback?: any;
  is_load_new_image: boolean;
  is_load_new_quiz: boolean;
  male_personality_type: string;
  male_character_type: string;
  female_personality_type: string;
  female_character_type: string;
  years_compatibility_str: string;
  description?: any;
};